Rename recval to librecval.
