+++ librecval/templates/list_phrases.html
